removed an ereg from config
added a qualifier to a str_replace for \t in Input

changed substr to strncmp in Codeigniter.php and directory_map function

added braces in an if statement of unit test

Removed "scripts" from the auto-load search path. Scripts were deprecated in Version 1.4.1 (September 21, 2006). If you still need to use them for legacy reasons, they must now be manually loaded in each Controller.

git-svn-id: http://dev.ellislab.com/svn/CodeIgniter/trunk@1133 c56aefdc-da1b-0410-bd20-cc00d14d2810